// src/lib/slack/types.ts
export type PlainText = {
	type: 'plain_text';
	text: string;
	emoji?: boolean;
};
export type MrkdwnText = {
	type: 'mrkdwn';
	text: string;
	verbatim?: boolean;
};
export type SlackText = PlainText | MrkdwnText;
export type ButtonElement = {
	type: 'button';
	text: PlainText;
	action_id?: string;
	url?: string;
	value?: string;
	style?: 'default' | 'primary' | 'danger';
	confirm?: ConfirmationDialog;
};
// ...
export type CheckboxElement = {
	type: 'checkboxes';
	action_id?: string;
	options: {
		text: MrkdwnText | PlainText;
		description?: MrkdwnText | PlainText;
		value: string;
	}[];
	initial_options?: {
		text: MrkdwnText | PlainText;
		value: string;
	}[];
	confirm?: ConfirmationDialog;
};
// ...
export type ActionElement =
	| ButtonElement
	| StaticSelectElement
	| DatePickerElement
	| OverflowElement
	| CheckboxElement
	| RadioButtonElement
	| TimePickerElement
	| DateTimePickerElement
	| MultiSelectElement
	| RichTextInputElement
	| WorkflowButtonElement;
export type HeaderBlock = {
	type: 'header';
	text: PlainText;
	block_id?: string;
};
export type SectionBlock = {
	type: 'section';
	text?: MrkdwnText | PlainText;
	fields?: (MrkdwnText | PlainText)[];
	accessory?: ActionElement;
	block_id?: string;
};
export type DividerBlock = {
	type: 'divider';
	block_id?: string;
};
export type ImageBlock = {
	type: 'image';
	image_url: string;
	alt_text: string;
	title?: PlainText;
	block_id?: string;
};
export type ActionsBlock = {
	type: 'actions';
	elements: ActionElement[];
	block_id?: string;
};
export type ContextBlock = {
	type: 'context';
	elements: (ImageElement | MrkdwnText | PlainText)[];
	block_id?: string;
};
export type InputBlock = {
	type: 'input';
	label: PlainText;
	element: ActionElement;
	dispatch_action?: boolean;
	block_id?: string;
	hint?: PlainText;
	optional?: boolean;
};
export type ImageElement = {
	type: 'image';
	image_url: string;
	alt_text: string;
};
export type FileBlock = {
	type: 'file';
	external_id: string;
	source: string;
	block_id?: string;
};
export type VideoBlock = {
	type: 'video';
	video_url: string;
	thumbnail_url: string;
	title: PlainText;
	alt_text: string;
	title_url?: string;
	description?: PlainText;
	provider_name?: string;
	provider_icon_url?: string;
	author_name?: string;
	block_id?: string;
};
export type ConfirmationDialog = {
	title?: PlainText;
	text: MrkdwnText | PlainText;
	confirm?: PlainText;
	deny?: PlainText;
	style?: 'primary' | 'danger';
};
export type InteractiveBlockyWocky =
	| HeaderBlock
	| SectionBlock
	| DividerBlock
	| ImageBlock
	| ActionsBlock
	| ContextBlock
	| InputBlock
	| FileBlock
	| VideoBlock;